Phase 03: CI Pipeline (#11)
* chore: add codebase map with parallel analysis

Generated .planning/codebase/ with 7 structured documents:
- STACK.md, INTEGRATIONS.md (tech stack & dependencies)
- ARCHITECTURE.md, STRUCTURE.md (design & file layout)
- CONVENTIONS.md, TESTING.md (code quality & test practices)
- CONCERNS.md (tech debt & improvement areas)

* docs: initialize project

* chore: add project config

* docs: complete project research

* docs: define v1 requirements

* docs: create roadmap (5 phases)

* docs(01): capture phase context

* docs(state): record phase 1 context session

* docs(01): research phase domain

* docs(phase-1): add validation strategy

* docs(01-foundation): create phase plan

* docs(license): add MIT LICENSE file

- MIT License with copyright (c) 2026 Amikos Tech Ltd.
- Establishes legal identity for open source release

* refactor(module): rename module path to github.com/amikos-tech/ami-gin

- Update go.mod module declaration from gin-index to ami-gin
- Update all 12 Go source file imports (cmd + examples)
- Update .golangci.yml gci prefix to match new module path
- Update README.md go get and import references
- Regenerate go.sum with new module path

* docs(01-01): complete license and module path plan

- Add 01-01-SUMMARY.md with execution results
- Update STATE.md with position, decisions, metrics
- Update ROADMAP.md with plan progress
- Mark FOUND-01, FOUND-02 requirements complete

* chore(01-02): remove internal PRD document

- Delete gin-index-prd.md (49KB internal planning doc with "Kiba Team" references)
- Prevents internal artifacts from leaking in public repository

* chore(01-02): add .planning/ to .gitignore and untrack

- Add .planning/ to .gitignore to prevent tracking GSD workflow artifacts
- Remove .planning/ from git tracking while preserving local files
- Ensures no internal planning artifacts leak in public repository

* fix(module): replace stale gin-index references in CLAUDE.md

Three auto-generated GSD blocks still referenced the old module path.

* chore: add .claude/settings.local.json to .gitignore

* chore: line-wrap LICENSE and remove empty Build Commands section

- Reformat LICENSE to canonical line-wrapped MIT format
- Remove duplicate empty Build Commands section from CLAUDE.md
  (real commands already in Build and Test Commands section above)

* feat(03-02): add SARIF security workflow

- add weekly and manual Security workflow
- upload govulncheck SARIF to code scanning

* chore(03-01): expand Makefile CI targets

- pin gotestsum installs to v1.13.0
- add integration-test and security-scan targets
- expose the full contributor command surface in help

* docs(03-02): add CI badge to README

- place the canonical ci.yml badge below the title
- keep README badge surface limited to GitHub-hosted CI

* docs(03-02): add plan 02 summary

- record task commits and verification evidence
- leave shared planning state files for orchestrator updates

* fix(03-01): expand lint policy and clear core findings

- enable the required CI linters in the v2 golangci config
- replace direct EOF comparisons with errors.Is handling
- remove unconvert and unparam blockers in tests and CLI parsing

* fix(03-01): resolve gosec findings in CLI and examples

- lock local index and sidecar writes down to 0600 permissions
- add explicit error handling across example document ingestion paths
- keep gosec focused on shipped code by excluding noisy G115 and test-only findings

* ci(03-01): add GitHub Actions CI workflow

* docs(03-01): add plan 01 summary

* ci(03): replace Codecov with GitHub-native coverage summary

* fix(03): WR-01 support regex predicates in CLI

* fix(03): WR-02 preserve local artifact permissions

* fix: harden predicate parsing and artifact modes

* test: harden parquet and predicate edge cases

* test: strengthen stale temp regression coverage

* test: polish review follow-ups

* ci: disable duplicate govulncheck checkout

+++ hyperloglog.go
@@ -77,7 +77,7 @@ func (hll *HyperLogLog) AddString(s string) {
 func (hll *HyperLogLog) addHash(hash uint64) {
 	idx := hash >> (64 - hll.precision)
 	w := hash<<hll.precision | (1 << (hll.precision - 1))
-	rho := uint8(leadingZeros(w) + 1)
+	rho := leadingZeros(w) + 1
 
 	if rho > hll.registers[idx] {
 		hll.registers[idx] = rho
